+++ src/views/changelog/ChangelogPage.tsx
@@ -82,7 +82,7 @@ const ChangelogPage: React.FC<Props> = () => {
           For more up-to-date and in-depth information, please check out the changelog in the <A href="https://github.com/tif-calin/votevote/blob/main/CHANGELOG.md">GitHub repo</A>. In addition, you can see <A href="https://dev--votevote.netlify.app/">this Netlify deployment</A> of the upcoming version.
         </p>
 
-        <h3>v0.1.3 <span>2022_05may??</span></h3>
+        {/* <h3>v0.1.3 <span>2022_05may??</span></h3>
         <div className="release">
           <h4>Added</h4>
           <ul>
@@ -99,7 +99,7 @@ const ChangelogPage: React.FC<Props> = () => {
             <li>a voter&apos;s input number amount updates on reset</li>
             <li>numerous browser styling inconsistencies now fixed</li>
           </ul>
-        </div>
+        </div> */}
 
         <br />
         <h3>v0.1.2 <span>2022_02feb28</span></h3>