+++ CHANGELOG.md
@@ -6,6 +6,14 @@ This changelog is inspired by [keepachangelog.com](https://keepachangelog.com/en
 ### Changed
  - first block's explanation defaults to open
  - add new preview img and apple icon for site links
+ - minor styling tweaks to left input and bar chart axis labels
+
+### Improved
+ - cross-browser styling tweaks
+    - firefox lacks support for invert() for backdrop-filter so I hardcoded the background color
+    - improved styling consistency for selects and inputs
+    - x-axis labels are the now in the same position
+    - firefox bar labels don't disappear when longer than bar
 
 ## [0.1.2.1] - 2022-03-01
 ### Fixed